moved props processing to RunConverter, added test profile
